split apart utils into thread manager and utils
