+++ README.md
@@ -172,10 +172,21 @@ Python dependencies are inside the requrements.txt
 - RaspberryPi
   - [ ] Add script to run backend on startup
 - Backend
-  - [x] Finish Serial Handler
-  - [x] Finish Database Handler
-  - [ ] Add Backend Handler
-  - [ ] Add Loadcell calibration and tare
+  - [x] Finish serial handler
+  - [x] Finish database handler
+  - [ ] Add backend handler thread
+  - [ ] Add loadcell calibration and tare
+  - [ ] Look into overhauling the threading system
+  - [ ] In the backend handler thread add an error
+        handling or a push error to db table.
+  - [ ] In the backend handler we could also have a status
+        table method which can keep track of threads life and
+        other system status information
+  - [ ] Overhaul logging system to allow for using different log levels,
+        switching log levels and having the logs possibly be handled by
+        one thread, maybe not even the backend thread but possibly a different
+        logging thread.
+  - [ ] Update DB URL.
 - Database
   - [ ] Add BackendCommands Table
 - Frontend
